classes in sepparedt files
